requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
python-telegram-bot==13.15
APScheduler==3.6.3
scikit-learn==1.5.0
pytest